MODIFIED ADAPTER TO SET ALL IN VIEWHOLDER
- Removed SubscriberItem class
- Modified SubscriberAdapter to set each view in ViewHolder class by
Edwin
